provide access to request when storing a token
By providing access to the request the token store is able to store tokens based on other indicators from the request, e.g. the ip address. This enables users to have more than one token, e.g. one for each device they are using.

In turn, it should not be required to find a token for a specific user. A login with the login provider should always yield a new token, old tokens must not be reused for security reasons.