Invalid request should trigger a 400 Bad Request, fixes #70
The uri extracted from request data (i.e. $_SERVER) might be invalid because
the issued request is erronous or the requestor tries to check if invalid
data is handled correctly. Until now such cases where detected with the
MalformedUriException, but as it was not catched it lead to a 500 Internal
Server Error response. With this fix the exception is catched, and a 400
Bad Request response is triggered instead.