Possibility to automatically generate and log request ids, fixes #74
With the new version 4.2 of stubbles/input the request provides access
to the request id send by the client, or generates one if none or an
invalid request id is present. The response adds this request id
as header when being send in case no specific request id header has
been set.